requirements.txt
pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.20.0
pytest>=7.0.0
ttkthemes>=3.2.0 